+++ PlanetRover/PlanetRover.Ui.ConsoleApplication/InputConverters/IDirectionConverter.cs
@@ -2,6 +2,6 @@ namespace PlanetRover.Ui.ConsoleApplication.InputConverters
 {
     interface IDirectionConverter : IInputConverter
     {
-        float Degree { get; }
+        int Degree { get; }
     }
 }
